socket io 기본 사용법실습

+++ controller/tweet.js
@@ -41,6 +41,13 @@ export async function updateTweet(req, res) {
 
 export async function deleteTweet(req, res) {
   const id = req.params.id;
+  const tweet = await tweetRepository.getById(id);
+  if (!tweet) {
+    return res.sendStatus(404);
+  }
+  if (tweet.userId !== req.userId) {
+    return res.sendStatus(403);
+  }
   await tweetRepository.remove(id);
   res.sendStatus(204);
 }
